Application layer restructured, classes separated and foldered
