+++ backend/MusicInteraction/MusicInteraction.Application/Commands/GradingMethodsCommands/GetGradingMethodByIdCommand.cs
@@ -1,0 +1,8 @@
+using MediatR;
+
+namespace MusicInteraction.Application;
+
+public class GetGradingMethodByIdCommand : IRequest<GetGradingMethodDetailResult>
+{
+    public Guid GradingMethodId { get; set; }
+}
