+++ backend/MusicInteraction/MusicInteraction.Application/Commands/GradingMethodsCommands/GetGradingMethodsCommand.cs
@@ -1,0 +1,5 @@
+using MediatR;
+
+namespace MusicInteraction.Application;
+
+public class GetPublicGradingMethodsCommand : IRequest<GetGradingMethodsResult> { }
