+++ backend/MusicInteraction/MusicInteraction.Application/Commands/InteractionsCommands/GetInteractionsCommand.cs
@@ -1,0 +1,6 @@
+using MediatR;
+
+namespace MusicInteraction.Application;
+
+public class GetInteractionsCommand: IRequest<GetInteractionsResult>
+{ }
