+++ backend/MusicInteraction/MusicInteraction.Application/Commands/InteractionsCommands/GetLikesCommand.cs
@@ -1,0 +1,5 @@
+using MediatR;
+
+namespace MusicInteraction.Application;
+
+public class GetLikesCommand: IRequest<GetLikesResult> { }
