+++ backend/MusicInteraction/MusicInteraction.Application/Commands/InteractionsCommands/GetRatingByIdCommand.cs
@@ -1,0 +1,8 @@
+using MediatR;
+
+namespace MusicInteraction.Application;
+
+public class GetRatingByIdCommand : IRequest<GetRatingDetailResult>
+{
+    public Guid RatingId { get; set; }
+}
