+++ backend/MusicInteraction/MusicInteraction.Application/Commands/InteractionsCommands/GetRatingsCommand.cs
@@ -1,0 +1,5 @@
+using MediatR;
+
+namespace MusicInteraction.Application;
+
+public class GetRatingsCommand: IRequest<GetRatingsResult> { }
