+++ backend/MusicInteraction/MusicInteraction.Application/Commands/InteractionsCommands/GetReviewsCommand.cs
@@ -1,0 +1,5 @@
+using MediatR;
+
+namespace MusicInteraction.Application;
+
+public class GetReviewsCommand: IRequest<GetReviewsResult> { }
