+++ backend/MusicInteraction/MusicInteraction.Application/DTOs/GradingMethodsDto/BlockComponentDto.cs
@@ -1,0 +1,7 @@
+namespace MusicInteraction.Application;
+
+public class BlockComponentDto : ComponentDto
+{
+    public List<ComponentDto> SubComponents { get; set; }
+    public List<Domain.Action> Actions { get; set; }
+}
