+++ backend/MusicInteraction/MusicInteraction.Application/DTOs/GradingMethodsDto/BlockDetailShowDto.cs
@@ -1,0 +1,7 @@
+namespace MusicInteraction.Application;
+
+public class BlockDetailShowDto : GradableComponentShowDto
+{
+    public List<GradableComponentShowDto> Components { get; set; }
+    public List<string> Actions { get; set; }
+}
