+++ backend/MusicInteraction/MusicInteraction.Application/DTOs/GradingMethodsDto/ComponentDto.cs
@@ -1,0 +1,11 @@
+using System.Text.Json.Serialization;
+
+namespace MusicInteraction.Application;
+
+[JsonPolymorphic(TypeDiscriminatorPropertyName = "componentType")]
+[JsonDerivedType(typeof(GradeComponentDto), typeDiscriminator: "grade")]
+[JsonDerivedType(typeof(BlockComponentDto), typeDiscriminator: "block")]
+public abstract class ComponentDto
+{
+    public string Name { get; set; }
+}
