+++ backend/MusicInteraction/MusicInteraction.Application/DTOs/GradingMethodsDto/GradeDetailShowDto.cs
@@ -1,0 +1,6 @@
+namespace MusicInteraction.Application;
+
+public class GradeDetailShowDto : GradableComponentShowDto
+{
+    public float StepAmount { get; set; }
+}
