+++ backend/MusicInteraction/MusicInteraction.Application/DTOs/GradingMethodsDto/GradingMethodSummaryDto.cs
@@ -1,0 +1,10 @@
+namespace MusicInteraction.Application;
+
+public class GradingMethodSummaryDto
+{
+    public Guid Id { get; set; }
+    public string Name { get; set; }
+    public string CreatorId { get; set; }
+    public DateTime CreatedAt { get; set; }
+    public bool IsPublic { get; set; }
+}
