+++ backend/MusicInteraction/MusicInteraction.Application/DTOs/InteractionsDto/GradesDTOs/GradeDetailDTO.cs
@@ -1,0 +1,6 @@
+namespace MusicInteraction.Application;
+
+public class GradeDetailDTO : GradedComponentDTO
+{
+    public float StepAmount { get; set; }
+}
