+++ backend/MusicInteraction/MusicInteraction.Application/DTOs/InteractionsDto/GradesDTOs/GradedBlockDetailDTO.cs
@@ -1,0 +1,7 @@
+namespace MusicInteraction.Application;
+
+public class GradedBlockDetailDTO : GradedComponentDTO
+{
+    public List<GradedComponentDTO> Components { get; set; }
+    public List<string> Actions { get; set; }
+}
