+++ backend/MusicInteraction/MusicInteraction.Application/DTOs/InteractionsDto/RatingDTOs/RatingNormalizedDTO.cs
@@ -1,0 +1,7 @@
+namespace MusicInteraction.Application;
+
+public class RatingNormalizedDTO
+{
+    public Guid RatingId { get; set; }
+    public float? NormalizedGrade { get; set; }
+}
