+++ backend/MusicInteraction/MusicInteraction.Application/DTOs/InteractionsDto/ReviewDTOs/ReviewDTO.cs
@@ -1,0 +1,7 @@
+namespace MusicInteraction.Application;
+
+public class ReviewDTO
+{
+    public Guid ReviewId { get; set; }
+    public string ReviewText { get; set; }
+}
