+++ backend/MusicInteraction/MusicInteraction.Application/Results/GradingMethodsResults/CreateGradingMethodResult.cs
@@ -1,0 +1,8 @@
+namespace MusicInteraction.Application;
+
+public class CreateGradingMethodResult
+{
+    public bool Success { get; set; }
+    public Guid? GradingMethodId { get; set; }
+    public string ErrorMessage { get; set; }
+}
