+++ backend/MusicInteraction/MusicInteraction.Application/Results/GradingMethodsResults/GetGradingMethodDetailResult.cs
@@ -1,0 +1,8 @@
+namespace MusicInteraction.Application;
+
+public class GetGradingMethodDetailResult
+{
+    public bool Success { get; set; }
+    public GradingMethodDetailDto GradingMethod { get; set; }
+    public string ErrorMessage { get; set; }
+}
