+++ backend/MusicInteraction/MusicInteraction.Application/Results/InteractionsResults/GetInteractionsResult.cs
@@ -1,0 +1,7 @@
+namespace MusicInteraction.Application;
+
+public class GetInteractionsResult
+{
+    public bool InteractionsEmpty { get; set; }
+    public List<InteractionAggregateShowDto> Interactions { get; set; }
+}
