+++ backend/MusicInteraction/MusicInteraction.Application/Results/InteractionsResults/GetLikesResult.cs
@@ -1,0 +1,9 @@
+using MusicInteraction.Domain;
+
+namespace MusicInteraction.Application;
+
+public class GetLikesResult
+{
+    public bool LikesEmpty { get; set; }
+    public List<Like> Likes { get; set; }
+}
