+++ backend/MusicInteraction/MusicInteraction.Application/Results/InteractionsResults/GetRatingDetailResult.cs
@@ -1,0 +1,8 @@
+namespace MusicInteraction.Application;
+
+public class GetRatingDetailResult
+{
+    public bool Success { get; set; }
+    public string ErrorMessage { get; set; }
+    public RatingDetailDTO Rating { get; set; }
+}
