+++ backend/MusicInteraction/MusicInteraction.Application/Results/InteractionsResults/GetRatingsResult.cs
@@ -1,0 +1,7 @@
+namespace MusicInteraction.Application;
+
+public class GetRatingsResult
+{
+    public bool RatingsEmpty { get; set; }
+    public List<RatingOverviewDTO> Ratings { get; set; }
+}
