+++ backend/MusicInteraction/MusicInteraction.Application/Results/InteractionsResults/GetReviewsResult.cs
@@ -1,0 +1,9 @@
+using MusicInteraction.Domain;
+
+namespace MusicInteraction.Application;
+
+public class GetReviewsResult
+{
+    public bool ReviewsEmpty { get; set; }
+    public List<Review> Reviews { get; set; }
+}
