+++ backend/MusicInteraction/MusicInteraction.Application/Storages/IInteractionStorage.cs
